backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
joblib==1.4.2
python-dotenv==1.0.1
requests==2.32.3
SpeechRecognition==3.10.1
gTTS==2.5.1
googletrans==4.0.0-rc1